# [testbed] Add batcher performance tests