--- a/testbed/stabilitytests/metric_test.go
+++ b/testbed/stabilitytests/metric_test.go
@@ -26,6 +26,7 @@ func TestStabilityMetricsOTLP(t *testing.T) {
 		contribPerfResultsSummary,
 		nil,
 		nil,
+		nil,
 	)
 }
 
@@ -40,6 +41,7 @@ func TestStabilityMetricsOpenCensus(t *testing.T) {
 			ResourceCheckPeriod: resourceCheckPeriod,
 		},
 		contribPerfResultsSummary,
+		nil,
 		nil,
 		nil,
 	)
@@ -58,6 +60,7 @@ func TestStabilityMetricsCarbon(t *testing.T) {
 		contribPerfResultsSummary,
 		nil,
 		nil,
+		nil,
 	)
 }
 
@@ -72,6 +75,7 @@ func TestStabilityMetricsSignalFx(t *testing.T) {
 			ResourceCheckPeriod: resourceCheckPeriod,
 		},
 		contribPerfResultsSummary,
+		nil,
 		nil,
 		nil,
 	)