--- a/testbed/stabilitytests/trace_test.go
+++ b/testbed/stabilitytests/trace_test.go
@@ -52,6 +52,7 @@ func TestStabilityTracesOpenCensus(t *testing.T) {
 		contribPerfResultsSummary,
 		processorsConfig,
 		nil,
+		nil,
 	)
 }
 
@@ -67,6 +68,7 @@ func TestStabilityTracesSAPM(t *testing.T) {
 		},
 		contribPerfResultsSummary,
 		processorsConfig,
+		nil,
 		nil,
 	)
 }
@@ -84,6 +86,7 @@ func TestStabilityTracesOTLP(t *testing.T) {
 		contribPerfResultsSummary,
 		processorsConfig,
 		nil,
+		nil,
 	)
 }
 
@@ -99,6 +102,7 @@ func TestStabilityTracesJaegerGRPC(t *testing.T) {
 		},
 		contribPerfResultsSummary,
 		processorsConfig,
+		nil,
 		nil,
 	)
 }
@@ -116,6 +120,7 @@ func TestStabilityTracesZipkin(t *testing.T) {
 		contribPerfResultsSummary,
 		processorsConfig,
 		nil,
+		nil,
 	)
 }
 
@@ -131,6 +136,7 @@ func TestStabilityTracesDatadog(t *testing.T) {
 		},
 		contribPerfResultsSummary,
 		processorsConfig,
+		nil,
 		nil,
 	)
 }