--- a/testbed/tests/metric_test.go
+++ b/testbed/tests/metric_test.go
@@ -88,6 +88,7 @@ func TestMetric10kDPS(t *testing.T) {
 				performanceResultsSummary,
 				nil,
 				nil,
+				nil,
 			)
 		})
 	}