--- a/testbed/tests/trace_test.go
+++ b/testbed/tests/trace_test.go
@@ -149,6 +149,7 @@ func TestTrace10kSPS(t *testing.T) {
 				performanceResultsSummary,
 				processors,
 				nil,
+				nil,
 			)
 		})
 	}