#### Description

Add basic batching benchmarks. The primary intent of these is to help verify that we're not introducing performance regressions with https://github.com/open-telemetry/opentelemetry-collector/issues/8122.

I've taken the the 10k DPS benchmark for logs, and ran it in different configurations:

* Batching either enabled or disabled
* In-memory queue enabled or disabled
* Using the batch processor at the start of the pipeline or the new exporter batcher
* All of this either with no processors or with some basic filtering and transformation

I've reduced the input batch size to 10 to better capture the effect of having no batching at the start of the pipeline on processor performance, which is one of the concerns with moving batching to the exporter.

For now, I'd like to get some comments on whether this is sufficient, or if I should expand my scope to other signal types and/or different parameters for the benchmarks.

#### Current results

Test                                    |Result|Duration|CPU Avg%|CPU Max%|RAM Avg MiB|RAM Max MiB|Sent Items|Received Items|
----------------------------------------|------|-------:|-------:|-------:|----------:|----------:|---------:|-------------:|
Log10kDPSNoProcessors/No_batching,_no_queue|PASS  |     15s|    20.7|    22.7|         73|        103|    150100|        150100|
Log10kDPSNoProcessors/No_batching,_queue|PASS  |     15s|    19.1|    21.7|         73|        103|    150100|        150100|
Log10kDPSNoProcessors/Batch_size_1000_with_batch_processor,_no_queue|PASS  |     15s|    13.1|    14.3|         77|        109|    150100|        150100|
Log10kDPSNoProcessors/Batch_size_1000_with_batch_processor,_queue|PASS  |     15s|    12.7|    14.0|         75|        108|    150100|        150100|
Log10kDPSNoProcessors/Batch_size_1000_with_exporter_batcher,_no_queue|PASS  |     15s|    15.5|    17.3|         72|        101|    150100|        150100|
Log10kDPSNoProcessors/Batch_size_1000_with_exporter_batcher,_queue|PASS  |     15s|    14.3|    15.7|         72|        102|    150100|        150100|
Log10kDPSWithProcessors/No_batching,_no_queue|PASS  |     15s|    21.5|    23.0|         72|        103|    150100|        150100|
Log10kDPSWithProcessors/No_batching,_queue|PASS  |     15s|    22.2|    23.0|         72|        102|    150100|        150100|
Log10kDPSWithProcessors/Batch_size_1000_with_batch_processor,_no_queue|PASS  |     15s|    22.5|    26.0|         75|        107|    150100|        150100|
Log10kDPSWithProcessors/Batch_size_1000_with_batch_processor,_queue|PASS  |     15s|    18.0|    19.7|         73|        104|    150100|        150100|
Log10kDPSWithProcessors/Batch_size_1000_with_exporter_batcher,_no_queue|PASS  |     16s|    18.3|    20.7|         75|        106|    150100|        150100|
Log10kDPSWithProcessors/Batch_size_1000_with_exporter_batcher,_queue|PASS  |     15s|    17.1|    17.7|         73|        102|    150100|        150100|

It looks like the new batcher is a bit less performant if the pipeline doesn't contain any processors, but is in fact faster if processors are present, which is surprising to me. But this does assuage our fears that we'd tank processor performance by moving batching to the end of the pipeline.

#### Link to tracking issue
Fixes https://github.com/open-telemetry/opentelemetry-collector/issues/10836