@sfc-gh-sili 2024-11-22
@swiatekm Thank you, Mikołaj! This is great.
Here are a couple of things that I'd be curious to know:
1. Sanity check that batching in exporter is about the same resource consumption as batching in processor, which this PR does already.
2. Like Joshua mentioned, I am too curious if not batching earlier causes higher resource usage for more expensive processing down in the pipeline.
(3. I'd be curious how size of the goroutine pool used by exporter would affects the performance. seems it's not something we've looked into so far)