@swiatekm 2024-12-03
Can we merge this PR as-is, and I'll add another benchmark in a follow-up? @MovieStoreGuy 

@jmacd @sfc-gh-sili I can see the argument that we could get better performance on some processing tasks because batch size can affect CPU cache saturation. I'm not 100% sure how to verify this though. Maybe spawn 100 transform processors doing the same thing?