@swiatekm 2024-12-06
Reran the benchmark on current main:

# Test PerformanceResults
Started: Fri, 06 Dec 2024 15:18:42 +0100

Test                                    |Result|Duration|CPU Avg%|CPU Max%|RAM Avg MiB|RAM Max MiB|Sent Items|Received Items|
----------------------------------------|------|-------:|-------:|-------:|----------:|----------:|---------:|-------------:|
Log10kDPSNoProcessors/No_batching,_no_queue|PASS  |     15s|    19.9|    25.7|         75|        104|    150100|        150100|
Log10kDPSNoProcessors/No_batching,_queue|PASS  |     15s|    16.4|    17.3|         71|        100|    150100|        150100|
Log10kDPSNoProcessors/Batch_size_1000_with_batch_processor,_no_queue|PASS  |     15s|    12.6|    13.7|         76|        108|    150100|        150100|
Log10kDPSNoProcessors/Batch_size_1000_with_batch_processor,_queue|PASS  |     15s|    11.9|    12.7|         74|        106|    150100|        150100|
Log10kDPSNoProcessors/Batch_size_1000_with_exporter_batcher,_no_queue|PASS  |     15s|    13.8|    15.3|         70|        100|    150100|        150100|
Log10kDPSNoProcessors/Batch_size_1000_with_exporter_batcher,_queue|PASS  |     15s|    13.2|    13.7|         72|        104|    150100|        150100|
Log10kDPSWithProcessors/No_batching,_no_queue|PASS  |     15s|    19.3|    20.3|         73|        103|    150100|        150100|
Log10kDPSWithProcessors/No_batching,_queue|PASS  |     15s|    19.6|    21.0|         73|        102|    150100|        150100|
Log10kDPSWithProcessors/Batch_size_1000_with_batch_processor,_no_queue|PASS  |     15s|    20.0|    21.7|         73|        102|    150100|        150100|
Log10kDPSWithProcessors/Batch_size_1000_with_batch_processor,_queue|PASS  |     15s|    19.8|    20.7|         73|        104|    150100|        150100|
Log10kDPSWithProcessors/Batch_size_1000_with_exporter_batcher,_no_queue|PASS  |     15s|    17.1|    19.7|         74|        104|    150100|        150100|
Log10kDPSWithProcessors/Batch_size_1000_with_exporter_batcher,_queue|PASS  |     15s|    17.7|    20.3|         74|        104|    150100|        150100|

Total duration: 182s

I also ran it with the `exporter.UsePullingBasedExporterQueueBatcher` feature gate introduced in https://github.com/open-telemetry/opentelemetry-collector/pull/11637 enabled:

Test                                    |Result|Duration|CPU Avg%|CPU Max%|RAM Avg MiB|RAM Max MiB|Sent Items|Received Items|
----------------------------------------|------|-------:|-------:|-------:|----------:|----------:|---------:|-------------:|
Log10kDPSNoProcessors/No_batching,_no_queue|PASS  |     15s|    16.8|    18.0|         74|        103|    150100|        150100|
Log10kDPSNoProcessors/No_batching,_queue|PASS  |     15s|    16.5|    16.7|         73|        103|    150100|        150100|
Log10kDPSNoProcessors/Batch_size_1000_with_batch_processor,_no_queue|PASS  |     15s|    11.5|    12.3|         77|        109|    150100|        150100|
Log10kDPSNoProcessors/Batch_size_1000_with_batch_processor,_queue|PASS  |     15s|    12.2|    13.0|         76|        109|    150100|        150100|
Log10kDPSNoProcessors/Batch_size_1000_with_exporter_batcher,_no_queue|PASS  |     15s|    11.7|    12.3|         73|        104|    150100|        150100|
Log10kDPSNoProcessors/Batch_size_1000_with_exporter_batcher,_queue|PASS  |     15s|    12.8|    14.0|         76|        107|    150100|        150100|
Log10kDPSWithProcessors/No_batching,_no_queue|PASS  |     15s|    25.5|    27.0|         73|        104|    150100|        150100|
Log10kDPSWithProcessors/No_batching,_queue|PASS  |     15s|    21.6|    23.7|         74|        106|    150100|        150100|
Log10kDPSWithProcessors/Batch_size_1000_with_batch_processor,_no_queue|PASS  |     15s|    21.0|    24.0|         73|        105|    150100|        150100|
Log10kDPSWithProcessors/Batch_size_1000_with_batch_processor,_queue|PASS  |     15s|    20.7|    21.0|         72|        103|    150100|        150100|
Log10kDPSWithProcessors/Batch_size_1000_with_exporter_batcher,_no_queue|PASS  |     15s|    16.0|    20.0|         77|        109|    150100|        150100|
Log10kDPSWithProcessors/Batch_size_1000_with_exporter_batcher,_queue|PASS  |     15s|    15.5|    16.0|         78|        112|    150100|        150100|